refactor: int 변환 시, Math.toIntExact 사용

+++ backend/src/main/java/todoktodok/backend/notification/application/dto/response/NotificationResponse.java
@@ -12,7 +12,7 @@ public record NotificationResponse(
             final List<NotificationItemResponse> notificationItemResponses
     ) {
         this(
-                unreadCount.intValue(),
+                Math.toIntExact(unreadCount),
                 notificationItemResponses
         );
     }
